requirements.txt
aiohttp
python-telegram-bot

ipython
pandas
tabulate
